@@ -1,6 +1,6 @@
 package org.kpmp.stateManager;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
@@ -10,9 +10,9 @@ import java.util.Date;
 import java.util.List;
 
 import org.bson.Document;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import org.springframework.data.mongodb.core.MongoTemplate;
@@ -29,14 +29,15 @@ public class CustomStateRepositoryTest {
 	private MongoTemplate mongoTemplate;
 	private CustomStateRepository customRepo;
 
-	@Before
+	@BeforeEach
 	public void setUp() throws Exception {
-		MockitoAnnotations.initMocks(this);
+		MockitoAnnotations.openMocks(this);
 		customRepo = new CustomStateRepository(stateRepository, mongoTemplate);
 	}
 
-	@After
+	@AfterEach
 	public void tearDown() throws Exception {
+        MockitoAnnotations.openMocks(this).close();
 		customRepo = null;
 	}
 
