@@ -1,13 +1,13 @@
 package org.kpmp.stateManager;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
@@ -20,16 +20,17 @@ public class NotificationHandlerTest {
 	private RestTemplate restTemplate;
 	private NotificationHandler handler;
 
-	@Before
+	@BeforeEach
 	public void setUp() throws Exception {
-		MockitoAnnotations.initMocks(this);
+		MockitoAnnotations.openMocks(this);
 		handler = new NotificationHandler(restTemplate);
 		ReflectionTestUtils.setField(handler, "notificationServiceHost", "host");
 		ReflectionTestUtils.setField(handler, "notificationEndpoint", "/endpoint");
 	}
 
-	@After
+	@AfterEach
 	public void tearDown() throws Exception {
+        MockitoAnnotations.openMocks(this).close();
 		handler = null;
 	}
 
