@@ -1,21 +1,21 @@
 package org.kpmp.stateManager;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
 public class StateChangeEventTest {
 
 	private StateChangeEvent event;
 
-	@Before
+	@BeforeEach
 	public void setUp() throws Exception {
 		event = new StateChangeEvent("origin", "packageId", "state", "codicil");
 	}
 
-	@After
+	@AfterEach
 	public void tearDown() throws Exception {
 		event = null;
 	}
