@@ -1,23 +1,25 @@
 package org.kpmp.stateManager;
 
-import static org.junit.Assert.assertEquals;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.Date;
 
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
 
 public class StateTest {
 
 	private State state;
 
-	@Before
+	@BeforeEach
 	public void setUp() throws Exception {
 		state = new State();
 	}
 
-	@After
+	@AfterEach
 	public void tearDown() throws Exception {
 		state = null;
 	}
